move find_peaks_1/2d from utils
