+++ alphadia/peakgroup/utils.py
@@ -3,6 +3,7 @@ import numpy as np
 from numba.extending import overload
 
 from alphadia.numba.fft import NumbaContextOnly
+from alphadia.utils import USE_NUMBA_CACHING
 
 
 def assemble_isotope_mz(mono_mz, charge, isotope_intensity):
@@ -37,3 +38,68 @@ def _(mono_mz, charge, isotope_intensity):
         return isotope_mz
 
     return funcx_impl
+
+
+@nb.njit(cache=USE_NUMBA_CACHING)
+def find_peaks_1d(a, top_n=3):
+    """accepts a dense representation and returns the top three peaks"""
+
+    scan = []
+    dia_cycle = []
+    intensity = []
+
+    for p in range(2, a.shape[1] - 2):
+        isotope_is_peak = (
+            a[0, p - 2] < a[0, p - 1] < a[0, p] > a[0, p + 1] > a[0, p + 2]
+        )
+
+        if isotope_is_peak:
+            intensity.append(a[0, p])
+            scan.append(0)
+            dia_cycle.append(p)
+
+    scan = np.array(scan)
+    dia_cycle = np.array(dia_cycle)
+    intensity = np.array(intensity)
+
+    idx = np.argsort(intensity)[::-1][:top_n]
+
+    scan = scan[idx]
+    dia_cycle = dia_cycle[idx]
+    intensity = intensity[idx]
+
+    return scan, dia_cycle, intensity
+
+
+@nb.njit(cache=USE_NUMBA_CACHING)
+def find_peaks_2d(a, top_n=3):
+    """accepts a dense representation and returns the top three peaks"""
+    scan = []
+    dia_cycle = []
+    intensity = []
+
+    for s in range(2, a.shape[0] - 2):
+        for p in range(2, a.shape[1] - 2):
+            isotope_is_peak = (
+                a[s - 2, p] < a[s - 1, p] < a[s, p] > a[s + 1, p] > a[s + 2, p]
+            )
+            isotope_is_peak &= (
+                a[s, p - 2] < a[s, p - 1] < a[s, p] > a[s, p + 1] > a[s, p + 2]
+            )
+
+            if isotope_is_peak:
+                intensity.append(a[s, p])
+                scan.append(s)
+                dia_cycle.append(p)
+
+    scan = np.array(scan)
+    dia_cycle = np.array(dia_cycle)
+    intensity = np.array(intensity)
+
+    idx = np.argsort(intensity)[::-1][:top_n]
+
+    scan = scan[idx]
+    dia_cycle = dia_cycle[idx]
+    intensity = intensity[idx]
+
+    return scan, dia_cycle, intensity
