+++ alphadia/utils.py
@@ -53,71 +53,6 @@ def candidate_hash(precursor_idx, rank):
 
 
 @nb.njit(cache=USE_NUMBA_CACHING)
-def find_peaks_1d(a, top_n=3):
-    """accepts a dense representation and returns the top three peaks"""
-
-    scan = []
-    dia_cycle = []
-    intensity = []
-
-    for p in range(2, a.shape[1] - 2):
-        isotope_is_peak = (
-            a[0, p - 2] < a[0, p - 1] < a[0, p] > a[0, p + 1] > a[0, p + 2]
-        )
-
-        if isotope_is_peak:
-            intensity.append(a[0, p])
-            scan.append(0)
-            dia_cycle.append(p)
-
-    scan = np.array(scan)
-    dia_cycle = np.array(dia_cycle)
-    intensity = np.array(intensity)
-
-    idx = np.argsort(intensity)[::-1][:top_n]
-
-    scan = scan[idx]
-    dia_cycle = dia_cycle[idx]
-    intensity = intensity[idx]
-
-    return scan, dia_cycle, intensity
-
-
-@nb.njit(cache=USE_NUMBA_CACHING)
-def find_peaks_2d(a, top_n=3):
-    """accepts a dense representation and returns the top three peaks"""
-    scan = []
-    dia_cycle = []
-    intensity = []
-
-    for s in range(2, a.shape[0] - 2):
-        for p in range(2, a.shape[1] - 2):
-            isotope_is_peak = (
-                a[s - 2, p] < a[s - 1, p] < a[s, p] > a[s + 1, p] > a[s + 2, p]
-            )
-            isotope_is_peak &= (
-                a[s, p - 2] < a[s, p - 1] < a[s, p] > a[s, p + 1] > a[s, p + 2]
-            )
-
-            if isotope_is_peak:
-                intensity.append(a[s, p])
-                scan.append(s)
-                dia_cycle.append(p)
-
-    scan = np.array(scan)
-    dia_cycle = np.array(dia_cycle)
-    intensity = np.array(intensity)
-
-    idx = np.argsort(intensity)[::-1][:top_n]
-
-    scan = scan[idx]
-    dia_cycle = dia_cycle[idx]
-    intensity = intensity[idx]
-
-    return scan, dia_cycle, intensity
-
-
-@nb.njit(cache=USE_NUMBA_CACHING)
 def amean1(array):
     out = np.zeros(array.shape[0])
     for i in range(len(out)):
